+++ lib/ja_serializer/builder/relationship.ex
@@ -3,6 +3,7 @@ defmodule JaSerializer.Builder.Relationship do
 
   alias JaSerializer.Builder.Link
   alias JaSerializer.Builder.ResourceIdentifier
+  alias JaSerializer.Builder.Utils
 
   defstruct [:name, :links, :data, :meta]
 
@@ -14,7 +15,10 @@ defmodule JaSerializer.Builder.Relationship do
         []
 
       _ ->
-        Enum.map(serializer.relationships(data, conn), &build(&1, context))
+        data
+        |> serializer.relationships(conn)
+        |> filter_fields(context)
+        |> Enum.map(&build(&1, context))
         |> Enum.filter(fn r -> not empty?(r) end)
     end
   end
@@ -26,6 +30,30 @@ defmodule JaSerializer.Builder.Relationship do
     |> add_links(definition, context)
     |> add_data(definition, context)
   end
+
+  defp filter_fields(
+         relationships,
+         context = %{serializer: serializer, opts: opts}
+       ) do
+    case opts[:fields] do
+      fields when is_map(fields) ->
+        do_filter(
+          relationships,
+          fields[serializer.type(context.data, context.conn)]
+        )
+
+      _any ->
+        relationships
+    end
+  end
+
+  defp do_filter(relationships, nil), do: relationships
+
+  defp do_filter(relationships, fields) when is_list(fields),
+    do: Map.take(relationships, fields)
+
+  defp do_filter(relationships, fields) when is_binary(fields),
+    do: do_filter(relationships, Utils.safe_atom_list(fields))
 
   defp empty?(%__MODULE__{data: nil, links: nil, meta: nil}), do: true
   defp empty?(%__MODULE__{} = _relationship), do: false